Refactor camera alarm timeouts
