+++ .homeycompose/drivers/settings/alarmTimeout.json
@@ -1,0 +1,12 @@
+{
+  "id": "alarm_timeout",
+  "type": "number",
+  "label": {
+    "en": "Alarm Timeout"
+  },
+  "value": 10,
+  "min": 1,
+  "units": {
+    "en": "seconds"
+  }
+}
